refactor(ai): split AIAssistantPanel into composer/transcript/streaming submodules (#5493)
Closes #5493

Splits the 1286-line AIAssistantPanel into focused sub-modules under
`src/shared/python/ai/gui/assistant/`:
- `composer.py`  — ChatInput widget and input-area builder
- `transcript.py` — MessageWidget and scrollable message area
- `streaming.py` — StreamWorker QThread
- `panel.py`     — AIAssistantPanel orchestration (lean at 708 lines)
- `__init__.py`  — public re-exports

Keeps `assistant_panel.py` as a thin backward-compatible shim (22 lines).
Removes the file-size budget exception added in PR #5466.
